feat(extension): add warning modal for paper wallet closure
