+++ apps/browser-extension-wallet/src/views/browser-view/features/settings/components/PaperWallet/index.ts
@@ -1,0 +1,4 @@
+export * from './PassphraseStage';
+export * from './SecureStage';
+export * from './SaveStage';
+export * from './types';
