+++ apps/browser-extension-wallet/src/views/browser-view/features/settings/components/PaperWallet/types.ts
@@ -1,0 +1,4 @@
+export interface Validation {
+  error?: string;
+  success?: string;
+}
