+++ apps/browser-extension-wallet/src/views/browser-view/features/settings/components/SettingsLayout.module.scss
@@ -201,6 +201,10 @@
   background-color: var(--data-pink) !important;
 }
 
+.settingsExitButton {
+  background-color: var(--primary-default);
+}
+
 .settingsSelect {
   width: 100%;
   margin-top: size_unit(6) !important;
