+++ packages/translation/src/lib/translations/browser-extension-wallet/en.json
@@ -846,5 +846,8 @@
   "paperWallet.enterPgpPrivateKey.fileUploadLabel": "Choose file to upload",
   "paperWallet.enterPgpPrivateKey.supportedFileUploadFormats": "Supported formats: .asc",
   "browserView.settings.generatePaperWallet.description": "Download an encrypted backup of your recovery phrase as a paper wallet.",
-  "paperWallet.WalletOverview.accountZero": "Account #0"
+  "paperWallet.WalletOverview.accountZero": "Account #0",
+  "paperWallet.SettingsDrawer.ExitWarningTitle": "Exit paper wallet setup?",
+  "paperWallet.SettingsDrawer.ExitWarningDescription": "You have generated a paper wallet, but haven't downloaded it yet! You'll have to generate a new one next time.",
+  "paperWallet.SettingsDrawer.ExitButton": "Exit"
 }
